refactor(roller): unify duplicated modifier schemas into a single source
Every modifier had its notation schema implemented twice: src/modifiers/<mod>.ts
(schema + behavior + docs, driving roll()) and src/notation/definitions/<mod>.ts
(schema-only, driving parse/validate/tokenize). The copies had drifted and the
main bundle shipped both.

- notation/definitions/<mod>.ts is now the single canonical source of each
  modifier's pattern/parse/toNotation/toDescription. Each modifiers/<mod>.ts
  imports its schema and attaches only apply/validate behavior.
- Reconciled the drift: compound/explode/penetrate toDescription now use the
  richer formatComparisonDescription form (the modifiers-copy behavior a test
  pinned); everything else was semantically identical.
- Moved docs to a dedicated pure-data module (src/docs/modifierDocData.ts) so
  dist/docs/index.js is self-contained with zero roll-engine/parse code and the
  tokenize path stays schema-only.
- Tightened size-limit budgets: dist/index.js 16 -> 13.25 KB, dist/docs 20 ->
  5.25 KB, dist/trace 5 -> 1.1 KB.

Public surface unchanged (RDN conformance vectors byte-identical). sideEffects
left true: setting it false triggers a pre-existing bunup/rolldown bug that drops
re-exported bindings from the barrel entry (reproducible on origin/main).
