+++ packages/roller/__tests__/docs/singleEntryModifierDocs.test.ts
@@ -1,48 +1,49 @@
 import { describe, expect, test } from 'bun:test'
-import { capSchema } from '../../src/modifiers/cap'
-import { replaceSchema } from '../../src/modifiers/replace'
-import { explodeSchema } from '../../src/modifiers/explode'
-import { compoundSchema } from '../../src/modifiers/compound'
-import { penetrateSchema } from '../../src/modifiers/penetrate'
-import { wildDieSchema } from '../../src/modifiers/wildDie'
-import { uniqueSchema } from '../../src/modifiers/unique'
-import { multiplySchema } from '../../src/modifiers/multiply'
-import { plusSchema } from '../../src/modifiers/plus'
-import { minusSchema } from '../../src/modifiers/minus'
-import { integerDivideSchema } from '../../src/modifiers/integerDivide'
-import { moduloSchema } from '../../src/modifiers/modulo'
-import { sortSchema } from '../../src/modifiers/sort'
-import { multiplyTotalSchema } from '../../src/modifiers/multiplyTotal'
+import {
+  capDocs,
+  compoundDocs,
+  explodeDocs,
+  integerDivideDocs,
+  minusDocs,
+  moduloDocs,
+  multiplyDocs,
+  multiplyTotalDocs,
+  penetrateDocs,
+  plusDocs,
+  replaceDocs,
+  sortDocs,
+  uniqueDocs,
+  wildDieDocs
+} from '../../src/docs/modifierDocData'
 
 describe('single-entry modifier docs (S3)', () => {
-  const singleEntrySchemas = [
-    { schema: capSchema, expectedKey: 'C{..}' },
-    { schema: replaceSchema, expectedKey: 'V{..}' },
-    { schema: explodeSchema, expectedKey: '!' },
-    { schema: compoundSchema, expectedKey: '!!' },
-    { schema: penetrateSchema, expectedKey: '!p' },
-    { schema: wildDieSchema, expectedKey: 'W' },
-    { schema: uniqueSchema, expectedKey: 'U' },
-    { schema: multiplySchema, expectedKey: '*' },
-    { schema: plusSchema, expectedKey: '+' },
-    { schema: minusSchema, expectedKey: '-' },
-    { schema: integerDivideSchema, expectedKey: '//' },
-    { schema: moduloSchema, expectedKey: '%' },
-    { schema: sortSchema, expectedKey: 'sort' },
-    { schema: multiplyTotalSchema, expectedKey: '**' }
+  const singleEntryDocs = [
+    { name: 'cap', docs: capDocs, expectedKey: 'C{..}' },
+    { name: 'replace', docs: replaceDocs, expectedKey: 'V{..}' },
+    { name: 'explode', docs: explodeDocs, expectedKey: '!' },
+    { name: 'compound', docs: compoundDocs, expectedKey: '!!' },
+    { name: 'penetrate', docs: penetrateDocs, expectedKey: '!p' },
+    { name: 'wildDie', docs: wildDieDocs, expectedKey: 'W' },
+    { name: 'unique', docs: uniqueDocs, expectedKey: 'U' },
+    { name: 'multiply', docs: multiplyDocs, expectedKey: '*' },
+    { name: 'plus', docs: plusDocs, expectedKey: '+' },
+    { name: 'minus', docs: minusDocs, expectedKey: '-' },
+    { name: 'integerDivide', docs: integerDivideDocs, expectedKey: '//' },
+    { name: 'modulo', docs: moduloDocs, expectedKey: '%' },
+    { name: 'sort', docs: sortDocs, expectedKey: 'sort' },
+    { name: 'multiplyTotal', docs: multiplyTotalDocs, expectedKey: '**' }
   ] as const
 
-  for (const { schema, expectedKey } of singleEntrySchemas) {
-    test(`${schema.name}Schema has a docs array with exactly 1 entry (key: '${expectedKey}')`, () => {
-      expect(schema.docs).toBeDefined()
-      expect(schema.docs!.length).toBe(1)
-      expect(schema.docs![0]!.key).toBe(expectedKey)
+  for (const { name, docs, expectedKey } of singleEntryDocs) {
+    test(`${name} has a docs array with exactly 1 entry (key: '${expectedKey}')`, () => {
+      expect(docs.length).toBe(1)
+      expect(docs[0]!.key).toBe(expectedKey)
     })
   }
 
-  test('each single-entry schema docs entry has required NotationDoc fields', () => {
-    for (const { schema } of singleEntrySchemas) {
-      const doc = schema.docs![0]!
+  test('each single-entry docs entry has required NotationDoc fields', () => {
+    for (const { docs } of singleEntryDocs) {
+      const doc = docs[0]!
       expect(typeof doc.key).toBe('string')
       expect(typeof doc.category).toBe('string')
       expect(typeof doc.title).toBe('string')
@@ -57,9 +58,9 @@ describe('single-entry modifier docs (S3)', () => {
     }
   })
 
-  test('no single-entry schema docs entry has a displayOptional field', () => {
-    for (const { schema } of singleEntrySchemas) {
-      const doc = schema.docs![0]!
+  test('no single-entry docs entry has a displayOptional field', () => {
+    for (const { docs } of singleEntryDocs) {
+      const doc = docs[0]!
       expect('displayOptional' in doc).toBe(false)
     }
   })
