+++ packages/roller/package.json
@@ -113,7 +113,7 @@
   "size-limit": [
     {
       "path": "dist/index.js",
-      "limit": "16 KB"
+      "limit": "13.25 KB"
     },
     {
       "path": "dist/index.d.ts",
@@ -125,11 +125,11 @@
     },
     {
       "path": "dist/docs/index.js",
-      "limit": "20 KB"
+      "limit": "5.25 KB"
     },
     {
       "path": "dist/trace/index.js",
-      "limit": "5 KB"
+      "limit": "1.1 KB"
     }
   ]
 }
