+++ packages/roller/src/modifiers/count.ts
@@ -1,205 +1,8 @@
 import type { CountOptions } from '../notation/types'
-import { parseComparisonNotation } from '../notation/comparison'
-import { defineNotationSchema } from '../notation/schema'
-import type { NotationSchema } from '../notation/schema'
+import { countSchema } from '../notation/definitions/count'
 import { ModifierError } from '../errors'
-import type { ModifierDefinition } from './schema'
 import { matchesComparison } from '../lib/comparison/matchesComparison'
-import type { NotationDoc } from '../docs/modifierDocs'
-
-const countPattern = /#\{((?:>=|<=|>|<|=)?\d+(?:,(?:>=|<=|>|<|=)?\d+)*)\}/
-
-export const countSchema: NotationSchema<CountOptions> = defineNotationSchema<CountOptions>({
-  name: 'count',
-  priority: 80,
-
-  pattern: countPattern,
-
-  parse: notation => {
-    const match = countPattern.exec(notation)
-    if (!match) return {}
-
-    const comparison = parseComparisonNotation(match[1] ?? '')
-    const hasAbove =
-      comparison.greaterThan !== undefined || comparison.greaterThanOrEqual !== undefined
-    const hasBelow = comparison.lessThan !== undefined || comparison.lessThanOrEqual !== undefined
-
-    const result: CountOptions = { ...comparison }
-
-    if (hasAbove && hasBelow) {
-      result.deduct = true
-    }
-
-    return { count: result }
-  },
-
-  docs: [
-    {
-      key: '#{..}',
-      category: 'Reinterpret',
-      color: '#60a5fa',
-      colorLight: '#2563eb',
-      title: 'Count',
-      description:
-        'Count dice matching comparison conditions instead of summing values. More powerful than S{}/F{} sugar.',
-      displayBase: '#{..}',
-      comparisons: [
-        { operator: '>=n', note: 'count dice showing n or more' },
-        { operator: '>n', note: 'count dice showing more than n' },
-        { operator: '<n', note: 'count dice showing less than n' },
-        { operator: '<=n', note: 'count dice showing n or less' },
-        { operator: '=n', note: 'count dice showing exactly n' }
-      ],
-      forms: [{ notation: '#{...}', note: 'Comma-separate multiple conditions' }],
-      examples: [
-        {
-          description: 'Count dice >= 7',
-          notation: '5d10#{>=7}',
-          options: { sides: 10, quantity: 5, modifiers: { count: { greaterThanOrEqual: 7 } } }
-        },
-        {
-          description: 'Count >3, deduct <1',
-          notation: '5d10#{>3,<1}',
-          options: {
-            sides: 10,
-            quantity: 5,
-            modifiers: { count: { greaterThan: 3, lessThan: 1, deduct: true } }
-          }
-        }
-      ]
-    },
-    {
-      key: 'S{..}',
-      category: 'Reinterpret',
-      color: '#3b82f6',
-      colorLight: '#1d4ed8',
-      title: 'Count Successes',
-      description:
-        'Count dice that meet a threshold instead of summing values \u2014 used in dice pool systems.',
-      displayBase: 'S{..}',
-      forms: [
-        { notation: 'S{n}', note: 'Single success threshold' },
-        { notation: 'S{n,b}', note: 'Threshold + botch threshold' }
-      ],
-      examples: [
-        {
-          description: 'Count dice that rolled 7 or higher',
-          notation: '5d10S{7}',
-          options: { sides: 10, quantity: 5, modifiers: { count: { greaterThanOrEqual: 7 } } }
-        },
-        {
-          description: 'Successes \u2265 7, subtract botches \u2264 1',
-          notation: '5d10S{7,1}',
-          options: {
-            sides: 10,
-            quantity: 5,
-            modifiers: { count: { greaterThanOrEqual: 7, lessThanOrEqual: 1, deduct: true } }
-          }
-        }
-      ]
-    },
-    {
-      key: 'F{..}',
-      category: 'Reinterpret',
-      color: '#93c5fd',
-      colorLight: '#3b82f6',
-      title: 'Count Failures',
-      description:
-        'Count dice at or below a threshold instead of summing values. Sugar for Count with lessThanOrEqual.',
-      displayBase: 'F{..}',
-      forms: [{ notation: 'F{N}', note: 'Count failures <= N' }],
-      examples: [
-        {
-          description: 'Count dice <= 3',
-          notation: '5d10F{3}',
-          options: { sides: 10, quantity: 5, modifiers: { count: { lessThanOrEqual: 3 } } }
-        }
-      ]
-    },
-    {
-      key: 'ms{..}',
-      category: 'Scale',
-      color: '#6366f1',
-      colorLight: '#4338ca',
-      title: 'Margin of Success',
-      description:
-        'Subtract a target number from the total to get the margin of success or failure. Sugar for Minus N.',
-      displayBase: 'ms{..}',
-      forms: [{ notation: 'ms{N}', note: 'Subtract N from total' }],
-      examples: [
-        {
-          description: 'Margin of success vs DC 15',
-          notation: '1d20ms{15}',
-          options: { sides: 20, modifiers: { minus: 15 } }
-        }
-      ]
-    }
-  ] satisfies readonly NotationDoc[],
-
-  toNotation: options => {
-    const parts: string[] = []
-
-    if (options.exact?.length) {
-      parts.push(...options.exact.map(n => `=${n}`))
-    }
-    if (options.greaterThanOrEqual !== undefined) {
-      parts.push(`>=${options.greaterThanOrEqual}`)
-    }
-    if (options.greaterThan !== undefined) {
-      parts.push(`>${options.greaterThan}`)
-    }
-    if (options.lessThanOrEqual !== undefined) {
-      parts.push(`<=${options.lessThanOrEqual}`)
-    }
-    if (options.lessThan !== undefined) {
-      parts.push(`<${options.lessThan}`)
-    }
-
-    if (parts.length === 0) return undefined
-
-    return `#{${parts.join(',')}}`
-  },
-
-  toDescription: options => {
-    const descriptions: string[] = []
-
-    if (options.exact?.length) {
-      descriptions.push(`Count dice equal to ${options.exact.join(', ')}`)
-    }
-
-    if (options.greaterThanOrEqual !== undefined) {
-      descriptions.push(`Count dice greater than or equal to ${options.greaterThanOrEqual}`)
-    }
-
-    if (options.greaterThan !== undefined) {
-      descriptions.push(`Count dice greater than ${options.greaterThan}`)
-    }
-
-    if (options.lessThanOrEqual !== undefined) {
-      if (options.deduct) {
-        descriptions.push(`deduct dice less than or equal to ${options.lessThanOrEqual}`)
-      } else {
-        descriptions.push(`Count dice less than or equal to ${options.lessThanOrEqual}`)
-      }
-    }
-
-    if (options.lessThan !== undefined) {
-      if (options.deduct) {
-        descriptions.push(`deduct dice less than ${options.lessThan}`)
-      } else {
-        descriptions.push(`Count dice less than ${options.lessThan}`)
-      }
-    }
-
-    if (descriptions.length === 0) return []
-
-    if (options.deduct && descriptions.length > 1) {
-      return [descriptions.join(', ')]
-    }
-
-    return descriptions
-  }
-})
+import type { ModifierDefinition } from './schema'
 
 export const countModifier: ModifierDefinition<CountOptions> = {
   ...countSchema,
