+++ packages/roller/src/modifiers/reroll.ts
@@ -1,205 +1,10 @@
 import type { RerollOptions } from '../notation/types'
-import {
-  formatComparisonNotation,
-  hasConditions,
-  parseComparisonNotation
-} from '../notation/comparison'
-import { formatHumanList } from '../notation/formatHumanList'
-import { defineNotationSchema } from '../notation/schema'
-import type { NotationSchema } from '../notation/schema'
+import { rerollSchema } from '../notation/definitions/reroll'
 import { ModifierError } from '../errors'
 import { matchesComparison, validateComparisonOptions } from '../lib/comparison'
 import { MAX_REROLL_ATTEMPTS } from '../lib/constants'
 import type { ModifierDefinition } from './schema'
 import { assertRollFn } from './schema'
-import type { NotationDoc } from '../docs/modifierDocs'
-
-const comparisonGroup = '((?:>=|<=|>|<|=)?\\d+(?:,(?:>=|<=|>|<|=)?\\d+)*)'
-const rerollPattern = new RegExp(
-  `[Rr][Oo]\\{${comparisonGroup}\\}|[Rr]\\{${comparisonGroup}\\}(\\d+)?`,
-  'g'
-)
-
-export const rerollSchema: NotationSchema<RerollOptions> = defineNotationSchema<RerollOptions>({
-  name: 'reroll',
-  priority: 40,
-
-  pattern: /[Rr]([Oo])?\{((?:>=|<=|>|<|=)?\d+(?:,(?:>=|<=|>|<|=)?\d+)*)\}(\d+)?/,
-
-  docs: [
-    {
-      key: 'R{..}',
-      category: 'Substitute',
-      color: '#f472b6',
-      colorLight: '#db2777',
-      title: 'Reroll',
-      description:
-        'Reroll dice that match a condition. The new result stands (may reroll again if still matching).',
-      displayBase: 'R{..}',
-      forms: [
-        {
-          notation: 'R{...}',
-          note: 'Reroll until result no longer matches'
-        },
-        { notation: 'R{...}(d)', note: 'Max d reroll attempts' }
-      ],
-      comparisons: [
-        { operator: 'n', note: 'reroll dice showing exactly n' },
-        { operator: '>n', note: 'reroll dice showing more than n' },
-        {
-          operator: '>=n',
-          note: 'reroll dice showing n or more'
-        },
-        {
-          operator: '<n',
-          note: 'reroll dice showing less than n'
-        },
-        {
-          operator: '<=n',
-          note: 'reroll dice showing n or less'
-        }
-      ],
-      examples: [
-        {
-          description: 'Reroll any 1s',
-          notation: '4d6R{1}',
-          options: { sides: 6, quantity: 4, modifiers: { reroll: { exact: [1] } } }
-        },
-        {
-          description: 'Reroll results under 3',
-          notation: '2d10R{<3}',
-          options: { sides: 10, quantity: 2, modifiers: { reroll: { lessThan: 3 } } }
-        },
-        {
-          description: 'Reroll under 3, max 2 attempts',
-          notation: '4d6R{<3}2',
-          options: { sides: 6, quantity: 4, modifiers: { reroll: { lessThan: 3, max: 2 } } }
-        }
-      ]
-    },
-    {
-      key: 'ro{..}',
-      category: 'Substitute',
-      color: '#f472b6',
-      colorLight: '#db2777',
-      title: 'Reroll Once',
-      description:
-        'Reroll dice matching a condition with a maximum of 1 attempt. Sugar for Reroll with max=1.',
-      displayBase: 'ro{..}',
-      forms: [
-        {
-          notation: 'ro{...}',
-          note: 'Reroll once if condition met'
-        }
-      ],
-      comparisons: [
-        { operator: 'n', note: 'reroll dice showing exactly n' },
-        { operator: '>n', note: 'reroll dice showing more than n' },
-        {
-          operator: '>=n',
-          note: 'reroll dice showing n or more'
-        },
-        {
-          operator: '<n',
-          note: 'reroll dice showing less than n'
-        },
-        {
-          operator: '<=n',
-          note: 'reroll dice showing n or less'
-        }
-      ],
-      examples: [
-        {
-          description: 'Reroll 1s once',
-          notation: '4d6ro{1}',
-          options: { sides: 6, quantity: 4, modifiers: { reroll: { exact: [1], max: 1 } } }
-        },
-        {
-          description: 'Reroll under 3 once',
-          notation: '2d10ro{<3}',
-          options: { sides: 10, quantity: 2, modifiers: { reroll: { lessThan: 3, max: 1 } } }
-        }
-      ]
-    }
-  ] satisfies readonly NotationDoc[],
-
-  parse: notation => {
-    const matches = Array.from(notation.matchAll(rerollPattern))
-    if (matches.length === 0) return {}
-
-    const reroll: RerollOptions = {}
-
-    for (const match of matches) {
-      const isRerollOnce = match[1] !== undefined
-      const conditions = isRerollOnce ? match[1] : match[2]
-      const maxCount = isRerollOnce ? 1 : match[3] ? Number(match[3]) : undefined
-
-      if (maxCount !== undefined) {
-        reroll.max = maxCount
-      }
-
-      if (conditions) {
-        const parsed = parseComparisonNotation(conditions)
-        if (parsed.greaterThan !== undefined) reroll.greaterThan = parsed.greaterThan
-        if (parsed.greaterThanOrEqual !== undefined)
-          reroll.greaterThanOrEqual = parsed.greaterThanOrEqual
-        if (parsed.lessThan !== undefined) reroll.lessThan = parsed.lessThan
-        if (parsed.lessThanOrEqual !== undefined) reroll.lessThanOrEqual = parsed.lessThanOrEqual
-        if (parsed.exact) reroll.exact = [...(reroll.exact ?? []), ...parsed.exact]
-      }
-    }
-
-    return hasConditions(reroll) || reroll.max !== undefined ? { reroll } : {}
-  },
-
-  toNotation: options => {
-    const parts = formatComparisonNotation(options)
-    if (!parts.length) return undefined
-
-    if (options.max === 1) {
-      return `ro{${parts.join(',')}}`
-    }
-
-    const maxSuffix = options.max ? `${options.max}` : ''
-    return `R{${parts.join(',')}}${maxSuffix}`
-  },
-
-  toDescription: options => {
-    const { exact, greaterThan, greaterThanOrEqual, lessThan, lessThanOrEqual, max } = options
-    const rerollList: string[] = []
-
-    if (exact) {
-      exact.forEach(roll => rerollList.push(`${roll}`))
-    }
-
-    const greaterLessList: string[] = []
-    if (greaterThanOrEqual !== undefined) {
-      greaterLessList.push(`greater than or equal to ${greaterThanOrEqual}`)
-    }
-    if (greaterThan !== undefined) {
-      greaterLessList.push(`greater than ${greaterThan}`)
-    }
-    if (lessThanOrEqual !== undefined) {
-      greaterLessList.push(`less than or equal to ${lessThanOrEqual}`)
-    }
-    if (lessThan !== undefined) {
-      greaterLessList.push(`less than ${lessThan}`)
-    }
-
-    const exactList = formatHumanList(rerollList.map(Number))
-    const greaterLess = greaterLessList.join(' and ')
-
-    const conditions = [exactList, greaterLess].filter(Boolean).join(', ')
-    if (!conditions) return []
-
-    if (max === 1) {
-      return [`Reroll once ${conditions}`]
-    }
-
-    const maxText = max !== undefined ? ` (up to ${max} times)` : ''
-    return [`Reroll ${conditions}${maxText}`]
-  }
-})
 
 function rerollSingle(
   roll: number,
